[REQ] Require at least Python 3.12 (#27)
Using the latest Python version fixes a problem that was introduced with common sub-expression elimination (#26).
For CSE, we need to hash nodes. 
Some nodes that are relevant for us contain `slice(...)` objects in their arguments.
But `slice` objects only got hashable in Python 3.12 (https://stackoverflow.com/a/76562346).
Therefore, I am upgrading the requirements to Python 3.12.
The library should still work well if we do not want to simplify compute graphs with nodes that contain `slice`s.

* [REQ] Require at least Python 3.12

* [REF] Make `flake8` happy